feat(web-app): url strategy redefinition
