+++ services/web-app/components/asset-catalog-item/asset-catalog-item.js
@@ -140,7 +140,7 @@ const AssetCatalogItem = ({ asset, showImage, isGrid = false, otherFrameworkCoun
     [styles['anchor--grid']]: isGrid
   })
 
-  const anchorHref = `/assets/${asset.params.library}/latest/${getSlug(asset.content)}`
+  const anchorHref = `/libraries/${asset.params.library}/latest/assets/${getSlug(asset.content)}`
 
   const renderGrid = () => (
     <Column as="li" md={4}>
