+++ services/web-app/components/hero/hero.module.scss
@@ -50,11 +50,11 @@ $hero-height: convert.rem(560px);
   */
 
 .assets::before {
-  background: url(../../pages/assets/index/images/assets-hero_mobile.png) no-repeat bottom right /
-    100% auto;
+  background: url(../../pages/index/images/assets-hero_mobile.png) no-repeat bottom right / 100%
+    auto;
 
   @include breakpoint.breakpoint(md) {
-    background: url(../../pages/assets/index/images/assets-hero.png) no-repeat top right / auto
+    background: url(../../pages/index/images/assets-hero.png) no-repeat top right / auto
       $hero-height;
   }
 }
