+++ services/web-app/pages/about-carbon/how-carbon-works.mdx
@@ -46,7 +46,7 @@ of who maintains them.
 Additionally, makers can quickly find assets that are compliant with their own product or team
 through local lenses.
 
-Head over to the [component catalog](/assets/components) to see the many assets shared between
+Head over to the [component catalog](/catalogs/components) to see the many assets shared between
 product and digital teams. Progress on the new platform can be found in our
 [roadmap](/about-carbon/platform-roadmap).
