+++ services/web-app/pages/developing/index.mdx
@@ -38,26 +38,26 @@ The Carbon Design System is built React first, with official support for Web Com
 support core parts of the system in vanilla JS, Angular, Vue, and Svelte—and we offer guidelines for
 other frameworks.
 
-Head over to [Libraries](/assets/libraries) to complete the installation tutorial for your framework
+Head over to [Libraries](/libraries) to complete the installation tutorial for your framework
 of choice.
 
 <CardGroup>
-  <MiniCard title="React tutorial" href="/assets/carbon-components-react">
+  <MiniCard title="React tutorial" href="/libraries/carbon-components-react">
     <MdxIcon name="react" />
   </MiniCard>
-  <MiniCard title="Web Component tutorial" href="/assets/carbon-components--web-components">
+  <MiniCard title="Web Component tutorial" href="/libraries/carbon-components--web-components">
     <MdxIcon name="webcomponents" />
   </MiniCard>
-  <MiniCard title="Angular tutorial" href="/assets/carbon-components-angular">
+  <MiniCard title="Angular tutorial" href="/libraries/carbon-components-angular">
     <MdxIcon name="angular" />
   </MiniCard>
-  <MiniCard title="Vue tutorial" href="/assets/carbon-components-vue">
+  <MiniCard title="Vue tutorial" href="/libraries/carbon-components-vue">
     <MdxIcon name="vue" />
   </MiniCard>
-  <MiniCard title="Svelte tutorial" href="h/assets/carbon-components-svelte">
+  <MiniCard title="Svelte tutorial" href="/libraries/carbon-components-svelte">
     <MdxIcon name="svelte" />
   </MiniCard>
-  <MiniCard title="Vanilla tutorial" href="/assets/carbon-components">
+  <MiniCard title="Vanilla tutorial" href="/libraries/carbon-components">
     <MdxIcon name="js" />
   </MiniCard>
 </CardGroup>
@@ -91,7 +91,7 @@ best practices along the way. Everyone who completes the tutorial will earn an I
 
 ### Explore additional libraries
 
-Product and digital teams at IBM have additional [Libraries](/assets/libraries) that are built upon
+Product and digital teams at IBM have additional [Libraries](/libraries) that are built upon
 the Carbon Design System, with components and styling developed for specific experiences.
 
 <CardGroup>
