+++ services/web-app/pages/elements/2x-grid/implementation.mdx
@@ -729,7 +729,7 @@ adding helpful uses cases as we find them.
 ## Incorporating the Shell
 
 We cover more about shell behavior on the 2x Grid Basics tab in the
-[Grid influencers](/assets/elements/2x-grid/overview/#grid-influencers) section. However, we've had
+[Grid influencers](/elements/2x-grid/overview/#grid-influencers) section. However, we've had
 a lot of questions about how the shell works in conjunction with our grid. That's why we recently
 incorporated the shell into our grid tooling as a Sketch template. We'd also like to show it in
 action here in a product.
@@ -1016,8 +1016,8 @@ offer several classes for specifying aspect ratio for given content. We use thes
 maintain consistency throughout our UI.
 
 We discuss aspect ratios and various grid behaviors more in depth on the 2x Grid Basics tab, both in
-the [Aspect ratios](/assets/elements/2x-grid/overview/#aspect-ratio) section and in the
-[Grid behaviors](/assets/elements/2x-grid/overview/#grid-behaviors) section.
+the [Aspect ratios](/elements/2x-grid/overview/#aspect-ratio) section and in the
+[Grid behaviors](/elements/2x-grid/overview/#grid-behaviors) section.
 
 <div className={clsx(styles['grid-demo'], styles['aspect-ratio-demo'])}>
   <Grid condensed>
