+++ services/web-app/pages/elements/accessibility/overview.mdx
@@ -124,7 +124,7 @@ affects about 4% of the world’s population.
 #### Designing for low vision
 
 Low vision users may be using a screen reader to preview your website or experience, so be sure to
-follow [keyboard guidelines](/assets/elements/accessibility/keyboard) to ensure the page is read to
+follow [keyboard guidelines](/elements/accessibility/keyboard) to ensure the page is read to
 the user in a logical order.
 
 ### Deaf and hard-of-hearing users
