+++ services/web-app/pages/elements/icons/contribute.mdx
@@ -28,12 +28,12 @@ Don't see the icon you need in the library? Make your own! Follow these guidelin
 consistency and proper formatting.
 
 - All icons should be unique and not redundant with any existing icons in the system. Search the
-  [library](/assets/elements/icons/library) for the keyword(s) associated with your proposed new
+  [library](/elements/icons/library) for the keyword(s) associated with your proposed new
   icon to ensure that it is not already represented.
 - All icons should adhere to the
   [IBM Design Language visual style](https://www.ibm.com/design/language/iconography/ui-icons/design).
 - All icons should comply with IBM
-  [accessibility standards](/assets/elements/accessibility/overview).
+  [accessibility standards](/elements/accessibility/overview).
 - All icons should be usable across all supported platforms and devices.
 - All icons should be understandable by a global audience of users, regardless of nationality or
   language.
