+++ services/web-app/pages/elements/pictograms/contribute.mdx
@@ -32,7 +32,7 @@ visual consistency and proper formatting.
 - All pictograms should adhere to the
   [IBM Design Language visual style](https://www.ibm.com/design/language/elements/pictograms/design).
 - All pictograms should comply with IBM
-  [accessibility standards](/assets/elements/accessibility/overview).
+  [accessibility standards](/elements/accessibility/overview).
 - All pictograms should be usable across all supported platforms and devices.
 - All pictograms should be understandable by a global audience of users, regardless of nationality
   or language.
