+++ services/web-app/pages/elements/themes/overview.mdx
@@ -40,7 +40,7 @@ variables, eliminating the need to modify individual components.
 
 ### Default theme
 
-Carbon provides four themes as shown in the [color usage](/assets/elements/color/usage) page. When
+Carbon provides four themes as shown in the [color usage](/elements/color/usage) page. When
 `carbon-components` is downloaded and installed, the components are preset to use the default
 (White) theme.
 
@@ -216,7 +216,7 @@ There are several token categories:
 
 Each theme is assigned universal color variables, which are determined by common roles and usage.
 This allows for uniform color application across themes while maintaining full styling flexibility.
-For more information, see the [color](/assets/elements/color/overview) page.
+For more information, see the [color](/elements/color/overview) page.
 
 ```scss
 $theme--white: (
@@ -316,7 +316,7 @@ $theme--white: (
 
 Use the spacing scale when building individual components. It includes small increments needed to
 create appropriate spatial relationships for detail-level designs. For more information, see the
-[spacing](/assets/elements/spacing/overview/) page.
+[spacing](/elements/spacing/overview/) page.
 
 ```scss
   $spacing-01: 0.125rem,
@@ -339,7 +339,7 @@ create appropriate spatial relationships for detail-level designs. For more info
 Typography has four categories of type styles (universal, productive, editorial, and additional)
 that can be customized through tokens. These tokens are used both within components and across
 layouts. Type tokens are determined by their role across the system. For more information, see the
-[typography](/assets/elements/typography/overview/) page.
+[typography](/elements/typography/overview/) page.
 
 ```scss
 // Universal
@@ -835,7 +835,7 @@ the same variables and roles, with only the value changing for each individual t
 <Row>
 <Column colLg={12}>
 
-![Default theme applied](/assets/elements/themes/themes-1.png)
+![Default theme applied](/elements/themes/themes-1.png)
 
 </Column>
 </Row>
