+++ services/web-app/pages/faq/index.mdx
@@ -106,7 +106,7 @@ implementations.
 ### Where do I find charts and data visualization?
 
 Carbon charts and their corresponding assets and documentation are now in the
-[Data visualization](/assets/data-visualization) section.
+[Data visualization](/collections/data-visualization) section.
 
 ### Where can I find documentation for earlier versions of Carbon?
