+++ services/web-app/pages/index/images/assets-hero.png
