+++ services/web-app/pages/index/images/assets-hero_mobile.png
