+++ services/web-app/pages/libraries/[host]/[org]/[repo]/[library]/[ref]/assets/index.module.scss
