+++ services/web-app/pages/libraries/[host]/[org]/[repo]/[library]/[ref]/index.js
@@ -64,7 +64,7 @@ const Library = ({ libraryData, params, navData }) => {
   const { sponsor } = libraryData.params
   const SponsorIcon = teams[sponsor] ? teams[sponsor].pictogram : Svg64Community
 
-  const assetsPath = `/assets/${params.library}/${params.ref}/library-assets`
+  const assetsPath = `/libraries/${params.library}/${params.ref}/assets`
 
   let externalDocsLink
   if (libraryData.content.externalDocsUrl) {
@@ -93,7 +93,7 @@ const Library = ({ libraryData, params, navData }) => {
             </div>
           }
           subTitle="Inherits"
-          href={`/assets/${library}/${version ?? ''}`}
+          href={`/libraries/${library}/${version ?? ''}`}
           actionIcon="arrowRight"
         >
           <Svg32Library />
