+++ services/web-app/stories/pages/homepage.stories.js
@@ -5,13 +5,13 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import Assets from '@/pages/assets/index.js'
+import HomePage from '@/pages/index.js'
 
 import { pageConfig, PageLayout } from '../../.storybook/layout'
 
 const stories = {
-  title: 'Platform/Pages/Assets',
-  component: Assets,
+  title: 'Platform/Pages/HomePage',
+  component: HomePage,
   ...pageConfig
 }
 
@@ -19,7 +19,7 @@ export default stories
 
 const Template = (args) => (
   <PageLayout args={args}>
-    <Assets />
+    <HomePage />
   </PageLayout>
 )
