+++ sonar-project.properties
@@ -24,5 +24,5 @@ sonar.coverage.exclusions=\
 # Patterns used to exclude some files from duplication report
 sonar.cpd.exclusions=\
   services/web-app/data/*,\
-  services/web-app/pages/assets/[host]/[org]/[repo]/[library]/[ref]/[asset].js,\
-  services/web-app/pages/assets/[host]/[org]/[repo]/[library]/[ref]/index.js
+  services/web-app/pages/libraries/[host]/[org]/[repo]/[library]/[ref]/assets/[asset]/index.js,\
+  services/web-app/pages/libraries/[host]/[org]/[repo]/[library]/[ref]/index.js
